+++ ChangeLog.md
@@ -1,6 +1,7 @@
 ## 0.14.0
 
 * Remove no-longer-relevant allow-newer flag (doesn't work with Setup.hs)
+* --no-rebuild-cabal
 
 ## 0.13.1